Добавил type hints , а так же создал директорий, в котором классы из main.py распределил по отдельным модулям
