+++ Validation_request task # 1/my_exeptions.py
@@ -1,0 +1,22 @@
+class UrlError(Exception):
+    def __init__(self, *args):
+        self.message = args[0] if args else None
+
+    def __str__(self):
+        return f"[Error INFO]: {self.message}"
+
+
+class MethodError(Exception):
+    def __init__(self, *args):
+        self.message = args[0] if args else None
+
+    def __str__(self):
+        return f"[Error INFO]: {self.message}"
+
+
+class ParamsError(Exception):
+    def __init__(self, *args):
+        self.message = args[0] if args else None
+
+    def __str__(self):
+        return f"[Error INFO]: {self.message}"
